[CSM-467] Move wallet user secret to module with plain user secret
