+++ auxx/src/Command/Run.hs
@@ -30,8 +30,8 @@ import           Pos.Crypto                 (emptyPassphrase, encToPublic,
 import           Pos.DB.Class               (MonadGState (..))
 import           Pos.Launcher.Configuration (HasConfigurations)
 import           Pos.Util.CompileInfo       (HasCompileInfo)
-import           Pos.Util.UserSecret        (readUserSecret, usKeys, usWallet, userSecret)
-import           Pos.Wallet.Web.Secret      (WalletUserSecret (..))
+import           Pos.Util.UserSecret        (WalletUserSecret (..), readUserSecret,
+                                             usKeys, usWallet, userSecret)
 
 import           Command.Help               (helpMessage)
 import qualified Command.Rollback           as Rollback
